feat(fonts): integrate variable font files, fix glyph_id u16 truncation
- Add variable Source Code Pro (wght axis, 212KB) from Adobe GitHub
- Variable Nunito Sans (opsz/wght/wdth/YTLC axes, 556KB) already in share/
- Update init to load variable fonts instead of static versions
- Fix compositor scene_render.rs glyph_id truncation: u8 → u16
- Change GlyphCache.get() signature from u8 to u16 to prevent truncation
- Add font_axes() to shaping rasterizer for variable font axis parsing
- 12 new tests: axis detection for both variable fonts (VAL-VARFONT-001/004)
- All 1,366 tests pass, QEMU boots and renders text with variable fonts

+++ system/services/compositor/scene_render.rs
@@ -238,9 +238,10 @@ fn render_node(
                 let mut cx = gx0;
 
                 for sg in shaped_glyphs {
-                    // Look up glyph in the cache by its ID (as ASCII byte).
+                    // Look up glyph in the cache by its full u16 ID.
                     // For monospace text, glyph_id == byte value (set by core).
-                    if let Some((glyph, coverage)) = cache.get(sg.glyph_id as u8) {
+                    // Using u16 avoids truncation that would break glyphs with ID > 255.
+                    if let Some((glyph, coverage)) = cache.get(sg.glyph_id) {
                         let px = cx + glyph.bearing_x;
                         let py = gy0 + (cache.ascent as i32 - glyph.bearing_y);
